requirements.txt
Django==3.0.8
djangorestframework==3.11.0
drf-yasg==1.17.1
